requirements_flask.txt
Flask
Werkzeug==2.3.7
torch
torchvision==0.15.2
Pillow==10.0.0
numpy==1.24.3
